test: 테스트 격리용 Service테스트 추가
