+++ backend/src/test/java/com/woowacourse/friendogly/footprint/service/FootprintCommandServiceTest.java
@@ -5,18 +5,15 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.woowacourse.friendogly.exception.FriendoglyException;
 import com.woowacourse.friendogly.footprint.dto.request.SaveFootprintRequest;
-import com.woowacourse.friendogly.footprint.repository.FootprintRepository;
 import com.woowacourse.friendogly.member.domain.Member;
-import com.woowacourse.friendogly.member.repository.MemberRepository;
-import org.junit.jupiter.api.AfterEach;
+import com.woowacourse.friendogly.support.ServiceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-@SpringBootTest
-class FootprintCommandServiceTest {
+
+class FootprintCommandServiceTest extends ServiceTest {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
@@ -24,26 +21,14 @@ class FootprintCommandServiceTest {
     @Autowired
     private FootprintCommandService footprintCommandService;
 
-    @Autowired
-    private MemberRepository memberRepository;
-
-    @Autowired
-    private FootprintRepository footprintRepository;
-
-    @AfterEach
-    void cleanUp() {
-        footprintRepository.deleteAll();
-        memberRepository.deleteAll();
-    }
-
     @DisplayName("발자국 저장")
     @Test
     void save() {
         Member member = memberRepository.save(
-            Member.builder()
-                .name("name")
-                .email("test@test.com")
-                .build()
+                Member.builder()
+                        .name("name")
+                        .email("test@test.com")
+                        .build()
         );
 
         footprintCommandService.save(new SaveFootprintRequest(member.getId(), 30.0, 30.0));
@@ -54,30 +39,30 @@ class FootprintCommandServiceTest {
     @Test
     void save_Fail_IllegalMemberId() {
         assertThatThrownBy(
-            () -> footprintCommandService.save(new SaveFootprintRequest(1L, 30.0, 30.0))
+                () -> footprintCommandService.save(new SaveFootprintRequest(1L, 30.0, 30.0))
         ).isInstanceOf(FriendoglyException.class)
-            .hasMessage("존재하지 않는 사용자 ID입니다.");
+                .hasMessage("존재하지 않는 사용자 ID입니다.");
     }
 
     @DisplayName("발자국 저장 실패 - 30초 전에 이미 발자국을 남긴 경우")
     @Test
     void save_Fail_TooOftenSave() {
         Member member = memberRepository.save(
-            Member.builder()
-                .name("name")
-                .email("test@test.com")
-                .build()
+                Member.builder()
+                        .name("name")
+                        .email("test@test.com")
+                        .build()
         );
 
         jdbcTemplate.update("""
-            INSERT INTO footprint (member_id, latitude, longitude, created_at, is_deleted)
-            VALUES
-            (?, 0.00000, 0.00000, TIMESTAMPADD(SECOND, -29, NOW()), FALSE)
-            """, member.getId());
+                INSERT INTO footprint (member_id, latitude, longitude, created_at, is_deleted)
+                VALUES
+                (?, 0.00000, 0.00000, TIMESTAMPADD(SECOND, -29, NOW()), FALSE)
+                """, member.getId());
 
         assertThatThrownBy(
-            () -> footprintCommandService.save(new SaveFootprintRequest(member.getId(), 30.0, 30.0))
+                () -> footprintCommandService.save(new SaveFootprintRequest(member.getId(), 30.0, 30.0))
         ).isInstanceOf(FriendoglyException.class)
-            .hasMessage("마지막 발자국을 찍은 뒤 30초가 경과되지 않았습니다.");
+                .hasMessage("마지막 발자국을 찍은 뒤 30초가 경과되지 않았습니다.");
     }
 }
