+++ backend/src/test/java/com/woowacourse/friendogly/footprint/service/FootprintQueryServiceTest.java
@@ -7,20 +7,16 @@ import com.woowacourse.friendogly.footprint.dto.request.FindNearFootprintRequest
 import com.woowacourse.friendogly.footprint.dto.request.SaveFootprintRequest;
 import com.woowacourse.friendogly.footprint.dto.response.FindMyLatestFootprintTimeResponse;
 import com.woowacourse.friendogly.footprint.dto.response.FindNearFootprintResponse;
-import com.woowacourse.friendogly.footprint.repository.FootprintRepository;
 import com.woowacourse.friendogly.member.domain.Member;
-import com.woowacourse.friendogly.member.repository.MemberRepository;
+import com.woowacourse.friendogly.support.ServiceTest;
 import java.time.LocalDateTime;
 import java.util.List;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-@SpringBootTest
-class FootprintQueryServiceTest {
+class FootprintQueryServiceTest extends ServiceTest {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
@@ -31,33 +27,22 @@ class FootprintQueryServiceTest {
     @Autowired
     private FootprintCommandService footprintCommandService;
 
-    @Autowired
-    private MemberRepository memberRepository;
-
-    @Autowired
-    private FootprintRepository footprintRepository;
-
-    @AfterEach
-    void cleanUp() {
-        footprintRepository.deleteAll();
-        memberRepository.deleteAll();
-    }
 
     @DisplayName("현재 위치 기준 1km 이내 발자국의 수가 1개일 때, 1개의 발자국만 조회된다.")
     @Test
     void findNear() {
         Member member1 = memberRepository.save(
-            Member.builder()
-                .name("name1")
-                .email("test@test.com")
-                .build()
+                Member.builder()
+                        .name("name1")
+                        .email("test@test.com")
+                        .build()
         );
 
         Member member2 = memberRepository.save(
-            Member.builder()
-                .name("name2")
-                .email("test@test.com")
-                .build()
+                Member.builder()
+                        .name("name2")
+                        .email("test@test.com")
+                        .build()
         );
 
         double nearLongitude = 0.008993216;
@@ -70,13 +55,13 @@ class FootprintQueryServiceTest {
         footprintCommandService.save(new SaveFootprintRequest(member2.getId(), 0.0, farLongitude));
 
         List<FindNearFootprintResponse> nearFootprints = footprintQueryService.findNear(
-            member1.getId(), new FindNearFootprintRequest(0.0, 0.0));
+                member1.getId(), new FindNearFootprintRequest(0.0, 0.0));
 
         assertAll(
-            () -> assertThat(nearFootprints).extracting(FindNearFootprintResponse::latitude)
-                .containsExactly(0.0),
-            () -> assertThat(nearFootprints).extracting(FindNearFootprintResponse::longitude)
-                .containsExactly(nearLongitude)
+                () -> assertThat(nearFootprints).extracting(FindNearFootprintResponse::latitude)
+                        .containsExactly(0.0),
+                () -> assertThat(nearFootprints).extracting(FindNearFootprintResponse::longitude)
+                        .containsExactly(nearLongitude)
         );
     }
 
@@ -84,28 +69,28 @@ class FootprintQueryServiceTest {
     @Test
     void findNear24Hours() {
         Member member = memberRepository.save(
-            Member.builder()
-                .name("name")
-                .email("test@test.com")
-                .build()
+                Member.builder()
+                        .name("name")
+                        .email("test@test.com")
+                        .build()
         );
 
         jdbcTemplate.update("""
-            INSERT INTO footprint (member_id, latitude, longitude, created_at, is_deleted)
-            VALUES
-            (?, 0.00000, 0.00000, TIMESTAMPADD(HOUR, -25, NOW()), FALSE),
-            (?, 0.00000, 0.00000, TIMESTAMPADD(HOUR, -23, NOW()), FALSE),
-            (?, 0.00000, 0.00000, TIMESTAMPADD(HOUR, -22, NOW()), FALSE);
-            """, member.getId(), member.getId(), member.getId());
+                INSERT INTO footprint (member_id, latitude, longitude, created_at, is_deleted)
+                VALUES
+                (?, 0.00000, 0.00000, TIMESTAMPADD(HOUR, -25, NOW()), FALSE),
+                (?, 0.00000, 0.00000, TIMESTAMPADD(HOUR, -23, NOW()), FALSE),
+                (?, 0.00000, 0.00000, TIMESTAMPADD(HOUR, -22, NOW()), FALSE);
+                """, member.getId(), member.getId(), member.getId());
 
         List<FindNearFootprintResponse> nearFootprints = footprintQueryService.findNear(
-            member.getId(), new FindNearFootprintRequest(0.0, 0.0));
+                member.getId(), new FindNearFootprintRequest(0.0, 0.0));
 
         assertAll(
-            () -> assertThat(nearFootprints).extracting(FindNearFootprintResponse::latitude)
-                .containsExactly(0.00000, 0.00000),
-            () -> assertThat(nearFootprints).extracting(FindNearFootprintResponse::longitude)
-                .containsExactly(0.00000, 0.00000)
+                () -> assertThat(nearFootprints).extracting(FindNearFootprintResponse::latitude)
+                        .containsExactly(0.00000, 0.00000),
+                () -> assertThat(nearFootprints).extracting(FindNearFootprintResponse::longitude)
+                        .containsExactly(0.00000, 0.00000)
         );
     }
 
@@ -113,28 +98,28 @@ class FootprintQueryServiceTest {
     @Test
     void findMyLatestFootprintTime_MyFootprintExists() {
         Member member = memberRepository.save(
-            Member.builder()
-                .name("name")
-                .email("test@test.com")
-                .build()
+                Member.builder()
+                        .name("name")
+                        .email("test@test.com")
+                        .build()
         );
 
         LocalDateTime oneMinuteAgo = LocalDateTime.now().minusMinutes(1);
 
         jdbcTemplate.update("""
-            INSERT INTO footprint (member_id, latitude, longitude, created_at, is_deleted)
-            VALUES
-            (?, 0.00000, 0.00000, TIMESTAMPADD(HOUR, -25, NOW()), FALSE),
-            (?, 0.11111, 0.11111, TIMESTAMPADD(HOUR, -23, NOW()), FALSE),
-            (?, 0.22222, 0.22222, ?, FALSE);
-            """, member.getId(), member.getId(), member.getId(), oneMinuteAgo);
+                INSERT INTO footprint (member_id, latitude, longitude, created_at, is_deleted)
+                VALUES
+                (?, 0.00000, 0.00000, TIMESTAMPADD(HOUR, -25, NOW()), FALSE),
+                (?, 0.11111, 0.11111, TIMESTAMPADD(HOUR, -23, NOW()), FALSE),
+                (?, 0.22222, 0.22222, ?, FALSE);
+                """, member.getId(), member.getId(), member.getId(), oneMinuteAgo);
 
         LocalDateTime time = footprintQueryService.findMyLatestFootprintTime(member.getId()).createdAt();
 
         assertAll(
-            () -> assertThat(time.getHour()).isEqualTo(oneMinuteAgo.getHour()),
-            () -> assertThat(time.getMinute()).isEqualTo(oneMinuteAgo.getMinute()),
-            () -> assertThat(time.getSecond()).isEqualTo(oneMinuteAgo.getSecond())
+                () -> assertThat(time.getHour()).isEqualTo(oneMinuteAgo.getHour()),
+                () -> assertThat(time.getMinute()).isEqualTo(oneMinuteAgo.getMinute()),
+                () -> assertThat(time.getSecond()).isEqualTo(oneMinuteAgo.getSecond())
         );
     }
 
@@ -142,14 +127,14 @@ class FootprintQueryServiceTest {
     @Test
     void findMyLatestFootprintTime_MyFootprintDoesNotExist() {
         Member member = memberRepository.save(
-            Member.builder()
-                .name("name")
-                .email("test@test.com")
-                .build()
+                Member.builder()
+                        .name("name")
+                        .email("test@test.com")
+                        .build()
         );
 
         assertThat(footprintQueryService.findMyLatestFootprintTime(member.getId()))
-            .extracting(FindMyLatestFootprintTimeResponse::createdAt)
-            .isNull();
+                .extracting(FindMyLatestFootprintTimeResponse::createdAt)
+                .isNull();
     }
 }
